gta view: add new command.
This is the beginning of an OpenGL-based viewer for array data.

Preliminary support is available for 2D arrays containing scalar values or
images. Planned features include support 3D volumes and other types of arrays.

As a side effect of this command, Qt4 support has been reenabled and will
remain the default until Qt5 is more solid.