libgta: deprecate compression for new version 1.2.0.
New files are always created without compression. Reading old files with
compression is only supported if the library is build with the
--with-compression configure option.

Building with CMake is now recommended over building with autoconf because CMake
module files will be generated and installed. Note that building with CMake only
supports the new default configuration with disabled compression.

In the future, both compression and autoconf support will be removed.